move response types to folder
