+++ src/types/rest/index.ts
@@ -1,19 +1,6 @@
-import { ConditionalOrderType, ConditionalOrderTypeNoUnderscore, OrderSide, OrderType } from "./orders";
+import { ConditionalOrderType, ConditionalOrderTypeNoUnderscore, OrderSide, OrderType } from "./response/orders";
 
-export * from "./api-key-info";
-export * from "./nft";
-export * from "./staking";
-export * from "./subaccount";
-export * from "./markets";
-export * from "./futures";
-export * from "./account";
-export * from "./wallet";
-export * from "./orders";
-export * from "./quote";
-export * from "./options";
-export * from "./spot-margin";
-export * from "./fills";
-export * from "./leveraged-tokens";
+export * from "./response";
 
 export interface TimeRange {
   start_time?: number;
