+++ src/types/rest/response/quote.ts
