+++ src/types/rest/response/spot-margin.ts
