+++ src/types/rest/response/staking.ts
